Website/time-series-master/requirements.txt
asgiref==3.2.5
beautifulsoup4==4.9.3
Brotli==1.0.9
certifi==2020.6.20
chardet==3.0.4
csscompressor==0.9.5
Django==3.0.4
django-compress-staticfiles==1.0.1b0
django-livereload-server==0.3.2
gunicorn==20.0.4
idna==2.10
numpy==1.18.2
pandas==1.0.3
python-dateutil==2.8.1
pytz==2019.3
requests==2.24.0
rjsmin==1.1.0
six==1.14.0
soupsieve==2.2.1
sqlparse==0.3.1
tornado==6.1
urllib3==1.25.10
